refactor: Refactor the asset command module
