+++ src-tauri/src/commands/mod.rs
@@ -1,18 +1,14 @@
 pub(crate) mod api_session;
-pub mod asset_detail;
-pub mod asset_favorite;
-pub mod asset_rename;
-pub mod asset_unfavorite;
-pub mod auth_flow;
-pub mod auth_logout;
-pub mod client_launcher;
-pub mod config_update;
-pub mod connect_methods;
-pub mod connect_token;
-pub mod dev_http_server;
-pub mod get_assets;
-pub mod get_config;
-pub mod get_setting;
-pub mod get_version;
-pub mod system_fonts;
-pub mod window_control;
+pub(crate) mod asset_actions;
+pub(crate) mod auth_flow;
+pub(crate) mod auth_logout;
+pub(crate) mod client_launcher;
+pub(crate) mod config_update;
+pub(crate) mod connect_methods;
+pub(crate) mod connect_token;
+pub(crate) mod dev_http_server;
+pub(crate) mod get_config;
+pub(crate) mod get_setting;
+pub(crate) mod get_version;
+pub(crate) mod system_fonts;
+pub(crate) mod window_control;
